test/requirements.txt
pytest==8.2.2
cocotb==1.9.1
cocotbext-spi==0.4.0
